Refactoring EmotionTracker
