+++ src/client/AM.PA.MacOs/PersonalAnalytics/Trackers/EmotionTracker/EmotionTrackerSettings.swift
@@ -1,0 +1,13 @@
+//
+//  EmotionTrackerSettings.swift
+//  PersonalAnalytics
+//
+//  Created by Roy Rutishauser on 20.11.19.
+//
+
+import Foundation
+
+
+enum EmotionTrackerSettings {
+    static let DbTable = "emotional_state"
+}
